requirements.txt
mediapipe==0.10.7
opencv-python-headless
Flask==3.0.0
pymongo==4.5.0
azure-core==1.29.4
azure-identity==1.14.1
azure-storage-blob==12.18.3